refactor: move entities to a separate entity package
Organized project structure by relocating all database entities to a dedicated entity package. This improves code maintainability and clarity.
