+++ src/main/java/app/cardcapture/sticker/dto/StickerResponseDto.java
@@ -1,6 +1,6 @@
 package app.cardcapture.sticker.dto;
 
-import app.cardcapture.sticker.domain.Sticker;
+import app.cardcapture.sticker.domain.entity.Sticker;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
