+++ src/main/java/app/cardcapture/template/dto/TemplateResponseDto.java
@@ -1,8 +1,8 @@
 package app.cardcapture.template.dto;
 
-import app.cardcapture.template.domain.Prompt;
-import app.cardcapture.template.domain.TemplateTag;
-import app.cardcapture.user.domain.User;
+import app.cardcapture.template.domain.entity.Prompt;
+import app.cardcapture.template.domain.entity.TemplateTag;
+import app.cardcapture.user.domain.entity.User;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
