+++ src/main/java/app/cardcapture/template/repository/PromptRepository.java
@@ -1,6 +1,6 @@
 package app.cardcapture.template.repository;
 
-import app.cardcapture.template.domain.Prompt;
+import app.cardcapture.template.domain.entity.Prompt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
