+++ src/main/java/app/cardcapture/template/repository/TemplateRepository.java
@@ -1,6 +1,6 @@
 package app.cardcapture.template.repository;
 
-import app.cardcapture.template.domain.Template;
+import app.cardcapture.template.domain.entity.Template;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
