+++ src/main/java/app/cardcapture/template/repository/TemplateTagRepository.java
@@ -1,7 +1,6 @@
 package app.cardcapture.template.repository;
 
-import app.cardcapture.sticker.domain.Sticker;
-import app.cardcapture.template.domain.TemplateTag;
+import app.cardcapture.template.domain.entity.TemplateTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
