+++ src/main/java/app/cardcapture/user/domain/entity/User.java
@@ -1,7 +1,22 @@
-package app.cardcapture.user.domain;
+package app.cardcapture.user.domain.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
 
 @Entity
 @Getter
@@ -29,13 +44,10 @@ public class User {
     @Column(nullable = false)
     private boolean verifiedEmail;
 
-    @Column(nullable = false)
     private String givenName;
 
-    @Column(nullable = false)
     private String familyName;
 
-    @Column(nullable = true)
     private String picture;
 
     @Column(nullable = false)
