+++ src/main/java/app/cardcapture/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package app.cardcapture.user.repository;
 
-import app.cardcapture.user.domain.User;
+import app.cardcapture.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
